Make alacritty start tmux session by default
By making alacritty start and attach to a tmux session this allows us to
create a dotfiles / notes directory window in the main session and have
the projects split off from the main session.

I've also renamed `dev-tmux` to `mux` as it's shorter and makes for a
nicer usage. It then doesn't conflict with the shopify `dev` tool when
tab completing.